refactor: use `unplugin-config` (#3106)
* refactor: use `unplugin-config`

* fix: preserve indentation in generated code
